launch dtr naturally
