+++ src/views/Patient/MedReqDropDown/cdsHooksCards/cdsHooksCard.tsx
@@ -6,7 +6,7 @@ import Client from 'fhirclient/lib/Client';
 
 import { Card as HooksCard, Link } from '../../../../cds-hooks/resources/HookTypes';
 import { SmartApp } from '../../../Questionnaire/SmartApp';
-import { AppContext } from '../../../Questionnaire/questionnaireUtil';
+import { AppContext, getAppContext } from '../../../Questionnaire/questionnaireUtil';
 
 // TODO:
 //  - create a css file for better style
@@ -104,28 +104,9 @@ const CdsHooksCard = (props: CdsHooksCardProps) => {
     } else if (link.type === 'smart') {
       console.log('    launch: ' + link.url.split('?')[0]);
       console.log(link);
-      const appContext: AppContext = {};
+      let appContext: AppContext = {};
       if (link.appContext) {
-        try {
-          // Fix + encoded spaces back to precent encoded spaces
-          const encodedAppString = link.appContext.replace(/\+/g, '%20');
-          const appString = decodeURIComponent(encodedAppString);
-          // Could switch to this later
-          appString.split('&').map(e => {
-            const temp = e.split('=');
-            if (
-              temp[0] === 'questionnaire' ||
-              temp[0] === 'order' ||
-              temp[0] === 'coverage' ||
-              temp[0] === 'response'
-            ) {
-              appContext[temp[0]] = temp[1];
-            }
-          });
-        } catch (e) {
-          console.log('failed to get appContext');
-          throw e;
-        }
+        appContext = getAppContext(link.appContext);
       }
       props.tabCallback(
         <SmartApp
